+++ paws.common/R/client.R
@@ -1,3 +1,4 @@
+#' @include config.R
 #' @include credentials.R
 #' @include handlers.R
 #' @include struct.R
@@ -126,24 +127,29 @@ set_sts_signing_region <- function(sts_regional_endpoint, region) {
 }
 
 # client_config returns a ClientConfig configured for the service.
-client_config <- function(service_name, endpoints, cfgs) {
+client_config <- function(service_name, endpoints, cfgs, service_id) {
   s <- new_session()
   if (!is.null(cfgs)) {
     s$config <- cfgs
   }
   # If region not defined, set it
   if (nchar(s$config$region) == 0) {
-    s$config$region <- get_region(cfgs$credentials$profile)
+    s$config$region <- get_region(cfgs[["credentials"]][["profile"]])
   }
   region <- s$config$region
   if (s$config$endpoint != "") {
     endpoint <- s$config$endpoint
     signing_region <- region
   } else {
-    sts_regional_endpoint <- s$config$sts_regional_endpoint
-    e <- resolver_endpoint(service_name, region, endpoints, sts_regional_endpoint)
-    endpoint <- e$endpoint
-    signing_region <- e$signing_region
+    endpoint <- get_service_endpoint(cfgs[["credentials"]][["profile"]], service_id)
+    if (!is.null(endpoint)) {
+      signing_region <- region
+    } else {
+      sts_regional_endpoint <- s$config$sts_regional_endpoint
+      e <- resolver_endpoint(service_name, region, endpoints, sts_regional_endpoint)
+      endpoint <- e$endpoint
+      signing_region <- e$signing_region
+    }
   }
   c <- ClientConfig(
     config = s$config,