+++ paws.common/NEWS.md
@@ -1,6 +1,7 @@
 # paws.common 0.6.1.9000
 * fix how `read_ini` reads empty profiles from ini files
 * fix is.atomic behavior for R v4.4+
+* support service specific endpoints (#667)
 
 # paws.common 0.6.1
 * support nested content within ini files (#667) [Configuration and credential file settings](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html#cli-configure-files-s3)